Fixed layout of the dialog box
