+++ src/smartpeak/includes.cmake
@@ -18,7 +18,6 @@ include(include/SmartPeak/core/RawDataProcessors/sources.cmake)
 include(include/SmartPeak/core/SequenceSegmentProcessors/sources.cmake)
 include(include/SmartPeak/core/SampleGroupProcessors/sources.cmake)
 include(include/SmartPeak/core/ApplicationProcessors/sources.cmake)
-include(include/SmartPeak/core/PresetWorkflows/sources.cmake)
 include(include/SmartPeak/cli/sources.cmake)
 include(include/SmartPeak/iface/sources.cmake)
 include(include/SmartPeak/algorithm/sources.cmake)
