+++ src/widgets/include/SmartPeak/PresetWorkflows/HPLC_UV_Unknowns.h
@@ -23,15 +23,15 @@
 
 #pragma once
 
-#include <SmartPeak/core/PresetWorkflows/PresetWorkflow.h>
+#include <SmartPeak/PresetWorkflows/PresetWorkflow.h>
 
 namespace SmartPeak
 {
-  struct FIAMS_Unknowns : public PresetWorkflow {
+  struct HPLC_UV_Unknowns : public PresetWorkflow {
 
     virtual std::string getName() const override 
     { 
-      return "FIAMS Unknowns";
+      return "HPLC UV Unknowns";
     };
 
     virtual std::vector<std::string> getWorkflowSteps() const override
@@ -39,18 +39,14 @@ namespace SmartPeak
       return
       {
         "LOAD_RAW_DATA",
-        "EXTRACT_SPECTRA_WINDOWS",
-        "MERGE_SPECTRA",
-        "PICK_2D_FEATURES",
-        "SEARCH_ACCURATE_MASS",
-        "STORE_ANNOTATIONS",
-        "STORE_FEATURES",
-        "ESTIMATE_FEATURE_BACKGROUND_INTERFERENCES",
-        "STORE_FEATURE_BACKGROUND_ESTIMATIONS",
-        "FILTER_FEATURES_BACKGROUND_INTERFERENCES",
-        "MERGE_FEATURES_MS1",
-        "MERGE_INJECTIONS",
-        "STORE_FEATURES_SAMPLE_GROUP"
+        "MAP_CHROMATOGRAMS",
+        "EXTRACT_CHROMATOGRAM_WINDOWS",
+        "ZERO_CHROMATOGRAM_BASELINE",
+        "PICK_MRM_FEATURES",
+        "QUANTIFY_FEATURES",
+        "CHECK_FEATURES",
+        "SELECT_FEATURES",
+        "STORE_FEATURES"
       };
     };
 
@@ -58,7 +54,7 @@ namespace SmartPeak
     {
       return
         "Lorem ipsum dolor sit amet, consectetur adipiscing elit. "
-        "Nunc pellentesque sagittis lorem non interdum. Donec pulvinar laoreet velit, in cursus sapien lacinia quis. " 
+        "Nunc pellentesque sagittis lorem non interdum. Donec pulvinar laoreet velit, in cursus sapien lacinia quis. "
         "Interdum et malesuada fames ac ante ipsum primis in faucibus. "
         "Cras tincidunt suscipit sem, id placerat elit lobortis a.";
     };
