+++ src/widgets/include/SmartPeak/PresetWorkflows/LCMS_DDA_Spectra_Library_Matching.h
@@ -23,7 +23,7 @@
 
 #pragma once
 
-#include <SmartPeak/core/PresetWorkflows/PresetWorkflow.h>
+#include <SmartPeak/PresetWorkflows/PresetWorkflow.h>
 
 namespace SmartPeak
 {
