+++ src/widgets/include/SmartPeak/PresetWorkflows/sources.cmake
